+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Note the the latest version is usually work in progress and may have not yet bee
 
 - All operations now run in top level transactions (fixes #87)
 - Add completions for `var update --new-env` (fixes #100)
+- Allow '' (empty string) as string flag value (with `warg` update) (fixes #101)
 
 # v0.0.18